adding Phenom[Internal]Utils.c/.h
These are new files to help improve the modularity
of coding in lalsimulation. In particular
for phenom waveforms.

src/LALSimIMRPhenomInternalUtils.c/.h

are designed so that you can include the .h file
in any C code in lalsimulation and be able to use the
functions defined in that file. However, these
functions are not SWIG wrapped.

src/LALSimIMRPhenomUtils.c/.h

are designed so that any XLAL SWIG wrapped functions
are defined here. In particular so that you can
also call the functions defined in here from python