remove bashism for setting --enable-nightly
for the past few nights the wheels have not been built with the
--enable-nightly flag, this was added using:

if [[ ${CI_PIPELINE_SOURCE} == schedule ]]; then EXTRA_CONFIG_FLAGS="--enable-nightly"; fi

and [[ ]] is a bash extension, as the default non-interactive shell in
Debian is /bin/dash, is this the source of the problem. Modify the test
to be POSIX complient.

refs #59

[ci skip]