add FIXME comments for removing deprecated builds
now that packages are built nightly and containers are created nightly
using these packages there is no need for the older containers built
from source, along with the nightly parallelised Stretch and EL7 builds
as the Deb and RPM builds test Stretch and EL7 respsectively.

mark these with FIXME comments for future removal.

[ci skip]