use -b option to dch for debian packages
the nightly build packages use a prerelease versioning scheme to ensure
that they are a lower version that release builds, this causes a problem
for testing release builds as the version used is less then the release
version and the packaging scripts fail. the use of the -b option to dch
ensure that this is only considered a warning.

this needs to be addressed in a better way but progress needs to be
made now.

[ci skip]