Work around incomplete PyPI release of cryptography
The last step of the wheel-building jobs is to test installing and
importing the wheels. The Python cryptogrpahy package is pulled in
as a dependency. A new version of cryptography was just released on
PyPI, but wheels have not yet been uploaded for all platforms. To
avoid having to build the lastest version of cryptography from
source, and to avoid needing to install headers and development
libraries for openssl, temporarily install the previously released
version of cryptography (2.5).