Alternative method for calculating reflective bound adjustment
This commit changes how proposals are adjusted to fall within the prior bounds
when the variable is linear, rather than circular. Instead of literally reflecting
back and forth, which can require many iterations if the proposed value is far
outside of the allowed range, the correct adjustment is now calculated by finding
the remainder when the excess of the proposed value from the nearest bound is
divided by that range.  This is similar to how circular variables are handled,
except now we must track whether the range divides the excess an even or odd
number of times.