Fix logic when looping over v3 failures
To prevent v3 from failing sometimes during PE, there was
a loop added around the waveform generator that perturbs the masses
and tries again. However, the logic of this loop is such that
if the wavefom generator fails in a correct way (in particular
with XLAL_EDOM when the waveform is too short) that error code
is not propogated to the PE code correctly and is instead
overwritte. This patch fixes this.