Fix an incosistency in lalinference when amporder is nonzero
For a time-domain, precessing EOB model the spins must always be
defined at the starting frequency. When amporder is non-zero,
the starting frequency is is different from the reference frequency.
Previosuly, the transformation from tilt angles to Cartesian spin
components was done at the reference frequency even though the
EOB waveform was started at a different frequency. Here we check
whether the template is a time domain precessing EOB model and if
so we esnure that the transformation function is called at the
starting frequency of the waveform.