dont silently disable libraries with missing deps
the top-level build was originally setup to automatically disable
libraries if any dependencies had been explicitly disabled. this is
causing confusion, this change removes to code that automatically
disables the libraries so that configure will exit with an error so it
is clear what the problem is and why the library will not be built.

closes: #200