Break lalsupport into smaller translation units to speed up build
This is a demonstration of a technique to speed up the build of lal
by breaking some very large source files into smaller translation
units. The lalsupport library contains several source files that
repeatedly expand macros to generate type-specialized versions of
functions. This is a severe bottleneck in the build because these
large files are built as single translation units.

This change, if applied throughout lalsupport, would dramatically
speed up build times with `make -j`.