[skip ci] Update XLALPrecessingNRSurDynamics interface
This was added as part of MR !1041, but was not reviewed as it was not
yet being used. This will be reviewed as part of the ongoing RemnantFits
review, but some API changes are required. I'm making the required
API changes as part of a separate MR so that this can be included
in the upcoming lalsimulation release and any future changes during the
review should be backwards compatible.