Fix Bh1-Bh2 swap bug in NRSur7dq4 when m1 < m2
When m1 < m2, the surrogate code previously did the following:
1. Swap the labels of the two BH, i.e. m1 <-> m2, chi1 <-> chi2.
2. Multiply the odd-m modes by -1 to undo this swap.

The assumption here is that the first step is only a rigid rotation, and
the second does undoes this rigid rotation.

However, the first step is only a rigid rotation if after the swap the
in-plane spins are also rotate by pi. This is now fixed.