Change the behaviour of the final spin fit
Under some very rare circumstances, the final spin formula may give
a negative number inside a square root. Previously this was handleded
by having a tolerance and just taking the abs of the argument of the
square root if it was below tolerance and emitting XLAL_EDOM otherwise.
Now, we follow what is done in the final spin python code and simply set
the argument inside the square root to 0 if it is <0. A warning is always
printed when this is done.

(cherry picked from commit 5d8209665aa249d621aff2d23aea72dffca860e1)