Change the behaviour of XLALSimInspiralFD for precessing EOB
Previosuly, this function would always start earlier than the
requested frequency in order to ensure that the waveform is
properly conditioned at the desired frequency. However,
for precessing EOB models this is not appropriate, because the
spins must be defined at the starting frequency of the simulation.

Thus for precessing, time-domain  models for which the reference
 frequency is the starting frequency the behaviour is changed
such that the waveforms are always generated at the requested
frequency.