add cosmic string kink & kink-kink waveforms
for cosmic string burst searches, in addition to the existing cusp search.
the kink waveform has a power-law index of -5/3 in the frequency domain, with a cutoff frequency.
the kink-kink waveform has a power-law index of -2 in the frequency domain, with no cutoff frequency.
LALSimBurst.c is majorly updated to model these waveform. Rest are I/O related.