lalpulsar/ConstructPLUT.c Add = to left border
This particular function deals with the case where the circles are
"pathological" (i.e. the stereographic projection produces circles with
huge radii which end up crossing some kind of bifurcation, changing
the location of their center from one side to the other).

In such case, one has to track whether the left border of the circle is
inside the patch or not; if it's not, then one has to add a column of
1s in order to mark the relevant "annuli" (remember: we track the
borders and then cumsum from left to right).

This patch covers the particular case where the left border of the
annulus touches one bin in the central region of the patch (i.e. there
are 1s at both sides of the touching point, being that point included).

This is the limit case of what we had before. As per some in-person discussion,
(cc @alicia.sintes, @rafel.jaume) no more changes have to be done.