lalinference: remove explicit gwpy dependency
this remove the gwpy dependency from the RPM and Debian packages as the
current versions of gwpy are only packaged under Conda. the codes that
utilise gwpy have import protections so if someone tries to use these
codes and don't have gwpy installed they will receive a message saying
that gwpy is required.