set an image for the docker jobs
Use docker:latest for the docker builds, this wasn't initially specified
as the default image for the runners was defined as docker:latest so we
didn't need to specify an image for docker builds and just use the
default. However for other configurations, such as a kubernates runner
at CNAF, using docker:latest as the default image doesn't make sense
therefore we can't assume that the default image is docker:latest.