lalpulsar: disable SIStr library support in RPMs
This disables usage of the SIStr library in the RPM packaging, used for
hardware injection support. This requires the use of cds-crtools and
this is not packaged for RL8, and there are no plans to make this
available. If the hardware injection codes are needed then the Conda
packages should be used.